Enemy Follows Player; Left click to fire; Started the finish of level 1
If the player gets in the area that is patrolled by the enemy, then the enemy follows the player. If the enemy touches the player, then the player loses health. You can now also fire by left clicking, along with the key F, however the shot does not go where the mouse is aimed. Starting to finish the level. Level 2 is going to be a maze that you have to get through before you die.